Changed lifecycle methods from public to open
My company is using PluggableApplicationDelegate in our iOS codebase. We'd like to time each ApplicationService's didFinishLaunching and didBecomeActive methods, to find where we can get performance gains. My initial thought was to subclass PluggableApplicationDelegate and add timers to the relevant methods, but because they're public (not open) this isn't possible.